incr: copy all the code in single file
